fix: :hammer: Add src/block/spawn.rs
Issue: #21
